requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.3
gspread==6.0.2
google-auth==2.28.1
python-dotenv==1.0.1
